Extact calculator into its own package.
